requirements.txt
scapy
treepoem
passwordgenerator
phonenumbers
requests
sockets
python-nmap
pyfiglet